% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ar1nv.R
\name{ar1nv}
\alias{ar1nv}
\title{AR1NV - Estimate the parameters for an AR(1) model}
\usage{
ar1nv(x)
}
\arguments{
\item{x}{One dimensional time series vector}
}
\value{
Return a list containing:
\item{g}{estimate of the lag-one autocorrelation.}
\item{a}{estimate of the noise variance.}
}
\description{
AR1NV - Estimate the parameters for an AR(1) model
}
\references{
SGrinsted, A., J. C. Moore, and S. Jevrejeva. 2004. Application of the cross
wavelet transform and wavelet coherence to geophysical time series.
\emph{Nonlinear Processes in Geophysics} 11:561-566.
}
\author{
Tunc Oygur (info@tuncoygur.com.tr)

Code based on a cross wavelet and wavelet coherence toolbox MATLAB package written by Eric Breitenberger
}
