% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/n.check.data.R
\name{n.check.datum}
\alias{n.check.datum}
\title{Helper function}
\usage{
n.check.datum(x)
}
\arguments{
\item{x}{matrix}
}
\value{
list(t, dt, n.obs)
}
\description{
Helper function
}
\note{
This function is not exported
}
